scripts/requirements.txt
python-dotenv==1.0.0
supabase==2.14.0
openai>=1.61.0
imaplib2==3.6
schedule==1.2.1
crewai==0.108.0
pydantic>=2.0.0